refactoring: fallbackCopying as separated utility function
