+++ src/components/ChatContent/AssistantInput.tsx
@@ -4,22 +4,7 @@ import { Markdown } from "../Markdown";
 import { Container, Box } from "@radix-ui/themes";
 import { ToolCall } from "../../services/refact";
 import { ToolContent } from "./ToolsContent";
-
-function fallbackCopying(text: string) {
-  const textArea = document.createElement("textarea");
-  textArea.value = text;
-
-  textArea.style.top = "0";
-  textArea.style.left = "0";
-  textArea.style.position = "fixed";
-
-  document.body.appendChild(textArea);
-  textArea.focus();
-  textArea.select();
-
-  document.execCommand("copy");
-  document.body.removeChild(textArea);
-}
+import { fallbackCopying } from "../../utils/fallbackCopying";
 
 type ChatInputProps = {
   message: string | null;
