+++ src/utils/copyChatHistoryToClipboard.ts
@@ -1,4 +1,5 @@
 import type { RootState } from "../app/store";
+import { fallbackCopying } from "./fallbackCopying";
 
 export const copyChatHistoryToClipboard = async (
   chatThread: RootState["history"]["thread"],
@@ -11,19 +12,3 @@ export const copyChatHistoryToClipboard = async (
     fallbackCopying(jsonString);
   }
 };
-
-function fallbackCopying(text: string) {
-  const textArea = document.createElement("textarea");
-  textArea.value = text;
-
-  textArea.style.top = "0";
-  textArea.style.left = "0";
-  textArea.style.position = "fixed";
-
-  document.body.appendChild(textArea);
-  textArea.focus();
-  textArea.select();
-
-  document.execCommand("copy");
-  document.body.removeChild(textArea);
-}
